Hand Tracking and Styling
Attempted to implement hand tracking by coverting the video stream to
hsv format and then trying to dect the contours of the hand so that a
bounding box could be created.

Attempted to make the app more presentable by adding css styling.

Added a form to indicate which sign is to be translated. Currently not
operational.